When testing std, don't create two variants of threading internals
