+++ src/libstd/sys_common/thread.rs
@@ -1,10 +1,11 @@
+#![allow(dead_code)]
+
 use boxed::FnBox;
 use env;
 use sync::atomic::{self, Ordering};
 use sys::stack_overflow;
 use sys::thread as imp;
 
-#[allow(dead_code)]
 pub unsafe fn start_thread(main: *mut u8) {
     // Next, set up our stack overflow handler which may get triggered if we run
     // out of stack.
