+++ src/tools/tidy/src/pal.rs
@@ -56,9 +56,11 @@ const EXCEPTION_PATHS: &[&str] = &[
     "src/libstd/tests/run-time-detect.rs" ,
     "src/libstd/sys_common/mod.rs",
     "src/libstd/sys_common/net.rs",
+    "src/libstd/thread/tests.rs",
+    "src/libstd/sync/mpsc", // some tests are only run on non-emscripten
+
     "src/libterm", // Not sure how to make this crate portable, but test crate needs it.
     "src/libtest", // Probably should defer to unstable `std::sys` APIs.
-    "src/libstd/sync/mpsc", // some tests are only run on non-emscripten
 
     // std testing crates, okay for now at least
     "src/libcore/tests",
